requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Document Parsing
pdfplumber==0.11.0

# AI/LLM Integration
google-generativeai==0.3.2

# Testing
pytest==7.4.4
hypothesis==6.98.3
httpx==0.26.0

# Environment Management
python-dotenv==1.0.0

# File Upload Support
python-multipart==0.0.6
